--HG--
branch : gushiku

+++ extensions/model-loaders/model-loaders/src/com/badlogic/gdx/graphics/g3d/loaders/ogre/mesh/Face.java
@@ -30,7 +30,7 @@ import javax.xml.bind.annotation.XmlType;
 @XmlRootElement(name = "face")
 public class Face {
 
-	@XmlAttribute(required = true) public int v1;
-	@XmlAttribute public int v2;
-	@XmlAttribute public int v3;
+	@XmlAttribute(required = true) public short v1;
+	@XmlAttribute public short v2;
+	@XmlAttribute public short v3;
 }
